+++ domain/src/main/java/tachiyomi/domain/manga/interactor/NetworkToLocalManga.kt
@@ -1,6 +1,7 @@
 package tachiyomi.domain.manga.interactor
 
 import tachiyomi.domain.manga.model.Manga
+import tachiyomi.domain.manga.model.toMangaUpdate
 import tachiyomi.domain.manga.repository.MangaRepository
 
 class NetworkToLocalManga(
@@ -14,6 +15,11 @@ class NetworkToLocalManga(
                 val id = insertManga(manga)
                 manga.copy(id = id!!)
             }
+            !localManga.favorite -> {
+                // if the manga isn't a favorite, update new info from source to db
+                manga.updateManga(localManga.id)
+                manga.copy(id = localManga.id)
+            }
             else -> {
                 localManga
             }
@@ -26,5 +32,15 @@ class NetworkToLocalManga(
 
     private suspend fun insertManga(manga: Manga): Long? {
         return mangaRepository.insert(manga)
+    }
+
+    private suspend fun Manga.updateManga(id: Long) {
+        mangaRepository.update(
+            toMangaUpdate()
+                .copy(
+                    id = id,
+                    thumbnailUrl = thumbnailUrl?.takeIf { it.isNotBlank() },
+                ),
+        )
     }
 }